requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==2.0.12
click==8.1.8
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.6.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-decouple==3.8
python-dotenv==1.1.0
pytz==2021.1
regex==2021.4.4
requests==2.26.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
ticktick-py==2.0.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.7
uvicorn==0.34.1
